requirements.txt
asgiref==3.7.2
boto3==1.34.46
botocore==1.34.46
Django==5.0.1
django-filter==23.5
django-phonenumber-field==7.3.0
django-storages==1.14.2
jmespath==1.0.1
phonenumbers==8.13.29
pillow==10.2.0
psycopg==3.1.17
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
